fix: Cleaned up code for user client and cluster parsing
